STM32H7: increase i2c slave rx limit.

+++ targets/TARGET_STM/TARGET_STM32L1/objects.h
@@ -111,8 +111,8 @@ struct i2c_s {
     volatile uint8_t pending_slave_tx_master_rx;
     volatile uint8_t pending_slave_rx_maxter_tx;
     uint8_t *slave_rx_buffer;
-    volatile uint8_t slave_rx_buffer_size;
-    volatile uint8_t slave_rx_count;
+    volatile uint16_t slave_rx_buffer_size;
+    volatile uint16_t slave_rx_count;
 #endif
 #if DEVICE_I2C_ASYNCH
     uint32_t address;
